requirements.txt
python-dotenv
gspread
oauth2client
python-telegram-bot